Exporting all domain data + Fixed issue of omf entry change on Export, seismic [devel]
Summary:
Using multizdomain values in UI. For simple file export the subselection field is changed from uiSeisSubsel to uiSeisTranf
Also while exporting using simple ascii file, the omf entry was overwritten. This changed the domain information to native suvey zdomain. This has now been fixed
OD-5344, OD-5348

Test Plan:
Tested by exporting a cropped volume of depth seismics. omf entry are not being changed
Both 2D and 3D case tested.
Simple and SEG-Y workflows tested

Reviewers: raman, arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com/D16847

+++ src/Seis/seisselection.cc
@@ -345,7 +345,7 @@ bool Seis::RangeSelData::setZRange( const Interval<float>& rg )
     tkzs_.zsamp_.stop = rg.stop;
     if ( rg.hasStep() )
     {
-	mDynamicCastGet(const StepInterval<int>*,zrg,&rg);
+	mDynamicCastGet(const StepInterval<float>*,zrg,&rg);
 	if ( zrg )
 	    tkzs_.zsamp_.step = zrg->step;
     }
